move the util to a better home
