+++ app/src/local-resources/instruments/hooks/index.ts
@@ -3,3 +3,4 @@ export * from './useHomePipettes'
 export * from './usePipetteModelSpecs'
 export * from './usePipetteNameSpecs'
 export * from './usePipetteSpecsv2'
+export * from './useTipAttachmentStatus'
