+++ app/src/local-resources/instruments/hooks/useTipAttachmentStatus/index.ts
