+++ app/src/organisms/Desktop/Devices/ProtocolRun/ProtocolRunHeader/RunHeaderModalContainer/hooks/useRunHeaderDropTip.ts
@@ -1,13 +1,9 @@
 import { useEffect } from 'react'
 
-import { useHost } from '@opentrons/react-api-client'
 import { RUN_STATUS_IDLE, RUN_STATUS_STOPPED } from '@opentrons/api-client'
 import { FLEX_ROBOT_TYPE, OT2_ROBOT_TYPE } from '@opentrons/shared-data'
 
-import {
-  useDropTipWizardFlows,
-  useTipAttachmentStatus,
-} from '/app/organisms/DropTipWizardFlows'
+import { useDropTipWizardFlows } from '/app/organisms/DropTipWizardFlows'
 import { useProtocolDropTipModal } from '../modals'
 import {
   useCloseCurrentRun,
@@ -16,13 +12,12 @@ import {
 } from '/app/resources/runs'
 import { isTerminalRunStatus } from '../../utils'
 import { lastRunCommandPromptedErrorRecovery } from '/app/local-resources/commands'
+import { useTipAttachmentStatus } from '/app/local-resources/instruments'
 
 import type { RobotType } from '@opentrons/shared-data'
 import type { Run, RunStatus } from '@opentrons/api-client'
-import type {
-  DropTipWizardFlowsProps,
-  PipetteWithTip,
-} from '/app/organisms/DropTipWizardFlows'
+import type { PipetteWithTip } from '/app/local-resources/instruments'
+import type { DropTipWizardFlowsProps } from '/app/organisms/DropTipWizardFlows'
 import type { UseProtocolDropTipModalResult } from '../modals'
 import type { PipetteDetails } from '/app/resources/maintenance_runs'
 
@@ -49,7 +44,6 @@ export function useRunHeaderDropTip({
   robotType,
   runStatus,
 }: UseRunHeaderDropTipParams): UseRunHeaderDropTipResult {
-  const host = useHost()
   const isRunCurrent = useIsRunCurrent(runId)
   const enteredER = runRecord?.data.hasEverEnteredErrorRecovery ?? false
 
@@ -66,7 +60,6 @@ export function useRunHeaderDropTip({
   } = useTipAttachmentStatus({
     runId,
     runRecord: runRecord ?? null,
-    host,
   })
 
   const dropTipModalUtils = useProtocolDropTipModal({
