+++ app/src/organisms/Desktop/Devices/ProtocolRun/ProtocolRunHeader/RunHeaderModalContainer/modals/ProtocolDropTipModal.tsx
@@ -20,8 +20,10 @@ import { useHomePipettes } from '/app/local-resources/instruments'
 
 import type { PipetteData } from '@opentrons/api-client'
 import type { IconProps } from '@opentrons/components'
-import type { UseHomePipettesProps } from '/app/local-resources/instruments'
-import type { TipAttachmentStatusResult } from '/app/organisms/DropTipWizardFlows'
+import type {
+  UseHomePipettesProps,
+  TipAttachmentStatusResult,
+} from '/app/local-resources/instruments'
 
 type UseProtocolDropTipModalProps = Pick<
   UseHomePipettesProps,
