+++ app/src/organisms/DropTipWizardFlows/hooks/index.ts
@@ -1,6 +1,5 @@
 export * from './errors'
 export * from './useDropTipWithType'
-export * from './useTipAttachmentStatus'
 export * from './useDropTipLocations'
 export { useDropTipRouting } from './useDropTipRouting'
 export { useDropTipWithType } from './useDropTipWithType'
