+++ app/src/organisms/DropTipWizardFlows/index.ts
@@ -1,6 +1,4 @@
 export * from './DropTipWizardFlows'
-export { useTipAttachmentStatus } from './hooks'
 export * from './TipsAttachedModal'
 
-export type { TipAttachmentStatusResult, PipetteWithTip } from './hooks'
 export type { FixitCommandTypeUtils } from './types'
