+++ app/src/organisms/ErrorRecoveryFlows/RecoveryOptions/ManageTips.tsx
@@ -24,10 +24,8 @@ import { DT_ROUTES } from '/app/organisms/DropTipWizardFlows/constants'
 import { SelectRecoveryOption } from './SelectRecoveryOption'
 
 import type { RecoveryContentProps, RecoveryRoute, RouteStep } from '../types'
-import type {
-  FixitCommandTypeUtils,
-  PipetteWithTip,
-} from '/app/organisms/DropTipWizardFlows'
+import type { FixitCommandTypeUtils } from '/app/organisms/DropTipWizardFlows'
+import type { PipetteWithTip } from '/app/local-resources/instruments'
 
 // The Drop Tip flow entry point. Includes entry from SelectRecoveryOption and CancelRun.
 export function ManageTips(props: RecoveryContentProps): JSX.Element {
