+++ app/src/organisms/ErrorRecoveryFlows/RecoveryOptions/SelectRecoveryOption.tsx
@@ -20,7 +20,7 @@ import {
 import { RecoverySingleColumnContentWrapper } from '../shared'
 
 import type { ErrorKind, RecoveryContentProps, RecoveryRoute } from '../types'
-import type { PipetteWithTip } from '/app/organisms/DropTipWizardFlows'
+import type { PipetteWithTip } from '/app/local-resources/instruments'
 
 // The "home" route within Error Recovery. When a user completes a non-terminal flow or presses "Go back" enough
 // to escape the boundaries of any route, they will be redirected here.
