+++ app/src/pages/ODD/RunSummary/index.tsx
@@ -67,15 +67,13 @@ import {
   EMPTY_TIMESTAMP,
   useCurrentRunCommands,
 } from '/app/resources/runs'
-import {
-  useTipAttachmentStatus,
-  handleTipsAttachedModal,
-} from '/app/organisms/DropTipWizardFlows'
+import { handleTipsAttachedModal } from '/app/organisms/DropTipWizardFlows'
 import { lastRunCommandPromptedErrorRecovery } from '/app/local-resources/commands'
+import { useTipAttachmentStatus } from '/app/local-resources/instruments'
 
 import type { IconName } from '@opentrons/components'
 import type { OnDeviceRouteParams } from '/app/App/types'
-import type { PipetteWithTip } from '/app/organisms/DropTipWizardFlows'
+import type { PipetteWithTip } from '/app/local-resources/instruments'
 
 export function RunSummary(): JSX.Element {
   const { runId } = useParams<
@@ -236,7 +234,6 @@ export function RunSummary(): JSX.Element {
   } = useTipAttachmentStatus({
     runId,
     runRecord: runRecord ?? null,
-    host,
   })
 
   // Determine tip status on initial render only. Error Recovery always handles tip status, so don't show it twice.
